Extracted InventoryGrid to a new file
